+++ app/Modules/Notification/Models/NotificationLog.php
@@ -5,6 +5,7 @@ namespace App\Modules\Notification\Models;
 use App\Modules\Notification\Enums\NotificationEvent;
 use App\Modules\User\Models\User;
 use App\Trait\DashboardModel;
+use App\Trait\Scopes\Searchable;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -26,6 +27,7 @@ use Illuminate\Support\Carbon;
 class NotificationLog extends Model
 {
     use DashboardModel;
+    use Searchable;
 
     public const SENT = 'sent';